Refactoring: SecurityContext
