+++ manager/api/security/pom.xml
@@ -24,6 +24,10 @@
       <groupId>${project.groupId}</groupId>
       <artifactId>apiman-manager-api-core</artifactId>
     </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>apiman-manager-api-rest-exceptions</artifactId>
+    </dependency>
 
     <!-- Third Party Dependencies -->
     <dependency>
